# src/main/resources/openapi/management.yml
openapi: 3.0.0
info:
  version: 2.0.0
  title: DXI
paths:
...
  /lecturer:
    get:
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            format: int64

      operationId: getAllLecturers
      description: Returns all lecturers
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LecturerFindResponse'
    post:
      operationId: createLecturer
      description: Create user lecturer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lecturer'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LecturerResponse'
...
  /lecturer/{id}:
    put:
      parameters:
        - name: id
          in: path
          schema:
            type: integer
            format: int64
          required: true
      operationId: updateLecturer
      description: Updates user lecturer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LecturerUpdate'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LecturerResponse'
    delete:
      parameters:
        - name: id
          in: path
          schema:
            type: integer
            format: int64
          required: true
      operationId: deleteLecturer
      description: Deletes user lecturer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /admin:
    get:
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            format: int64
      operationId: getAllAdmins
      description: Returns all admins
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminFindResponse'
    post:
      operationId: createAdmin
      description: Create user Admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Admin'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
...
  /admin/{id}:
    put:
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      operationId: updateAdmin
      description: Updates user admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdate'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      operationId: deleteAdmin
      description: Deletes user Admin
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
...
  /group/{semesterId}:
    parameters:
      - name: semesterId
        in: path
        schema:
          type: integer
          format: int64
        required: true
    get:
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            format: int64
      operationId: getSemesterGroups
      description: gets all groups related to semester
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupFindResponse'
  /group/{semesterId}/{topicId}/{maxSize}:
    post:
      parameters:
        - name: semesterId
          in: path
          schema:
            type: integer
            format: int64
          required: true
        - name: topicId
          in: path
          schema:
            type: integer
            format: int64
          required: true
        - name: maxSize
          in: path
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 10
          required: true
      operationId: addNewGroup
      description: Adds new Group
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
  /group/addStudents:
    post:
      parameters:
        - name: groupId
          in: query
          schema:
            type: integer
            format: int64
          required: true
        - name: studentIds
          in: query
          schema:
            type: array
            items:
              type: string
          required: true
      operationId: insertStudents
      description: inserts students into group
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
  /group/signUp/{groupId}:
    post:
      parameters:
        - name: groupId
          in: path
          schema:
            type: integer
            format: int64
          required: true
      operationId: signUpForGroup
      description: signs user for group
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
  /group/signUser/{groupId}/{userId}:
    post:
      parameters:
        - name: groupId
          in: path
          schema:
            type: integer
            format: int64
          required: true
        - name: userId
          in: path
          schema:
            type: string
          required: true

      operationId: signUserForGroup
      description: signs user for group
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
components:
  schemas:
...
    Lecturer:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        subject:
          type: string
        academicDegree:
          type: string
        cathedral:
          type: string
    LecturerUpdate:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        subject:
          type: string
        academicDegree:
          type: string
        cathedral:
          type: string
    LecturerResponse:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        subject:
          type: string
        academicDegree:
          type: string
        cathedral:
          type: string
        isActive:
          type: boolean
    LecturerFindResponse:
      type: object
      properties:
        startElement:
          type: integer
          format: int64
        totalCount:
          type: integer
          format: int64
        count:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/LecturerResponse'
    UserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        email:
          type: string
        firstName:
          type: string
        isActive:
          type: boolean
...
    Admin:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    AdminResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    AdminFindResponse:
      type: object
      properties:
        startElement:
          type: integer
          format: int64
        totalCount:
          type: integer
          format: int64
        count:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/AdminResponse'
    AdminUpdate:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
...
    StudentResponse:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        isActive:
          type: boolean
...
    GroupResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maxSize:
          type: integer
          format: int32
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentResponse'
    GroupFindResponse:
      type: object
      properties:
        startElement:
          type: integer
          format: int64
        totalCount:
          type: integer
          format: int64
        count:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/GroupResponse'
...
    Role:
      type: string
      enum:
        - ADMIN
        - LECTURER
        - STUDENT
        - USER